swarm: christmas edition package

+++ Makefile
@@ -64,14 +64,16 @@ release-tarball: $(RELEASE_TARBALL_FILENAME)
 # daemon/contrib/tarballs.  To use a custom tarballs cache directory,
 # export the TARBALLS environment variable.
 tarballs.manifest:
-	rm -rf daemon/contrib/native
-	mkdir -p daemon/contrib/native && \
-	cd daemon/contrib/native && \
-	../bootstrap && \
-        $(MAKE) list && \
-        $(MAKE) fetch -j && \
-	$(MAKE) --silent list-tarballs > $(CURDIR)/$@
-	rm -rf daemon/contrib/native
+	if [[ "${DISTRIBUTION}" != "appimage" ]]; then
+		rm -rf daemon/contrib/native
+		mkdir -p daemon/contrib/native && \
+		cd daemon/contrib/native && \
+		../bootstrap && \
+			$(MAKE) list && \
+			$(MAKE) fetch -j && \
+		$(MAKE) --silent list-tarballs > $(CURDIR)/$@
+		rm -rf daemon/contrib/native
+	fi
 
 # Generate the release tarball.  Note: to avoid building 1+ GiB
 # tarball containing all the bundled libraries, only the required
@@ -81,23 +83,25 @@ tarballs.manifest:
 # it should be built in a minimal container.)
 $(RELEASE_TARBALL_FILENAME): tarballs.manifest
 # Prepare the sources of the top repository and relevant submodules.
-	rm -f "$@"
-	mkdir $(TMPDIR)/ring-project
-	git archive HEAD | tar xf - -C $(TMPDIR)/ring-project
-	for m in daemon lrc client-gnome client-qt; do \
-		(cd "$$m" && git archive --prefix "$$m/" HEAD \
-			| tar xf - -C $(TMPDIR)/ring-project); \
-	done
-# Create the base archive.
-	tar --create --file $(TMPDIR)/ring-project.tar $(TMPDIR)/ring-project \
-		--transform 's,.*/ring-project,ring-project,'
-# Append the cached tarballs listed in the manifest.
-	tar --append --file $(TMPDIR)/ring-project.tar \
-		--files-from $< \
-		--transform 's,^.*/,ring-project/daemon/contrib/tarballs/,'
-	gzip $(TMPDIR)/ring-project.tar
-	mv $(TMPDIR)/ring-project.tar.gz "$@"
-	rm -rf $(TMPDIR)
+	if [[ "${DISTRIBUTION}" != "appimage" ]]; then
+		rm -f "$@"
+		mkdir $(TMPDIR)/ring-project
+		git archive HEAD | tar xf - -C $(TMPDIR)/ring-project
+		for m in daemon lrc client-gnome client-qt; do \
+			(cd "$$m" && git archive --prefix "$$m/" HEAD \
+				| tar xf - -C $(TMPDIR)/ring-project); \
+		done
+	# Create the base archive.
+		tar --create --file $(TMPDIR)/ring-project.tar $(TMPDIR)/ring-project \
+			--transform 's,.*/ring-project,ring-project,'
+	# Append the cached tarballs listed in the manifest.
+		tar --append --file $(TMPDIR)/ring-project.tar \
+			--files-from $< \
+			--transform 's,^.*/,ring-project/daemon/contrib/tarballs/,'
+		gzip $(TMPDIR)/ring-project.tar
+		mv $(TMPDIR)/ring-project.tar.gz "$@"
+		rm -rf $(TMPDIR)
+	fi
 
 #######################
 ## Packaging targets ##
